+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0 --- Infectivity Crisis (WIP)
 
 - Refactor configuration to enable multiple fitness functions
+- Change `fitness_model` in configuration to `host_model` and `host_fitness` respectively
 - Add `Hill` utility function for trait fitness
 - Add `HostSpecs` wrapper to access host implementation by index